refactor: useEffect 분리 및 post/delete/patch 기능 모듈화
